+++ source/isaaclab_tasks/isaaclab_tasks/contrib/cartpole_showcase/cartpole_camera/cartpole_camera_env.py
@@ -9,11 +9,12 @@ import gymnasium as gym
 import torch
 
 from isaaclab_tasks.core.cartpole.cartpole_direct_camera_env import CartpoleCameraEnv
-from isaaclab_tasks.core.cartpole.cartpole_direct_camera_env_cfg import CartpoleRGBCameraEnvCfg
+
+from .cartpole_camera_env_cfg import CartpoleCameraEnvCfg
 
 
 class CartpoleCameraShowcaseEnv(CartpoleCameraEnv):
-    cfg: CartpoleRGBCameraEnvCfg
+    cfg: CartpoleCameraEnvCfg
 
     def _pre_physics_step(self, actions: torch.Tensor) -> None:
         self.actions = actions.clone()
@@ -40,7 +41,7 @@ class CartpoleCameraShowcaseEnv(CartpoleCameraEnv):
             raise NotImplementedError(f"Action space {type(self.single_action_space)} not implemented")
 
         # set target
-        self._cartpole.set_joint_effort_target_index(target=target, joint_ids=self._cart_dof_idx)
+        self.cartpole.set_joint_effort_target_index(target=target, joint_ids=self._cart_dof_idx)
 
     def _get_observations(self) -> dict:
         # get camera data